backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
psycopg[binary]>=3.2.0
alembic>=1.13.0
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
slowapi>=0.1.9
python-multipart>=0.0.9
apscheduler>=3.10.4
bcrypt==4.0.1